Improve error handling.
Move away from `.expect(...)` and `.unwrap()`, as they will eventually
lead to a whole mess, where it's impossible to properly handle error
conditions.

Rather than using the `failure` or `error_chain` crates, we just
implement error conversion ourselves. It's really not that big a deal,
and it's nice to be able to see what's happening explicitly.